+++ RateBlog/Models/YoutubeData.cs
@@ -16,7 +16,7 @@ namespace Bestfluence.Models
         public virtual YoutubeAgeGroup YoutubeAgeGroup { get; set; }
         public virtual ICollection<YoutubeCountry> YoutubeCountry { get; set; }
 
-        public int Engagement { get; set; }
+        public double Engagement { get; set; }
         public int Views { get; set; }
         public double MaleViews { get; set; }
         public double FemaleViews { get; set; }
